+++ crates/core/src/flow/link_error.rs
@@ -1,6 +1,5 @@
 use super::locate_error::ProtoLocateError;
 use crate::layout::ProtoLayoutError;
-use crate::tool_error::ProtoToolError;
 use starbase_styles::{Style, Stylize};
 use starbase_utils::fs::FsError;
 use starbase_utils::json::JsonError;
@@ -23,10 +22,6 @@ pub enum ProtoLinkError {
     #[diagnostic(transparent)]
     #[error(transparent)]
     Locate(#[from] Box<ProtoLocateError>),
-
-    #[diagnostic(transparent)]
-    #[error(transparent)]
-    Tool(#[from] Box<ProtoToolError>),
 
     #[diagnostic(code(proto::link::failed_args_parse))]
     #[error("Failed to parse shim arguments string {}.", .args.style(Style::Shell))]
@@ -58,11 +53,5 @@ impl From<ProtoLayoutError> for ProtoLinkError {
 impl From<ProtoLocateError> for ProtoLinkError {
     fn from(e: ProtoLocateError) -> ProtoLinkError {
         ProtoLinkError::Locate(Box::new(e))
-    }
-}
-
-impl From<ProtoToolError> for ProtoLinkError {
-    fn from(e: ProtoToolError) -> ProtoLinkError {
-        ProtoLinkError::Tool(Box::new(e))
     }
 }
