+++ crates/core/src/flow/resolve_error.rs
@@ -1,16 +1,31 @@
 use crate::config_error::ProtoConfigError;
 use crate::layout::ProtoLayoutError;
-use crate::tool_error::ProtoToolError;
+use crate::tool_spec::Backend;
+use crate::utils::archive::ProtoArchiveError;
+use crate::utils::process::ProtoProcessError;
 use starbase_styles::{Style, Stylize};
+use starbase_utils::fs::FsError;
 use std::path::PathBuf;
 use thiserror::Error;
-use warpgate::WarpgatePluginError;
+use warpgate::{WarpgateClientError, WarpgatePluginError};
 
 #[derive(Error, Debug, miette::Diagnostic)]
 pub enum ProtoResolveError {
     #[diagnostic(transparent)]
     #[error(transparent)]
+    Archive(#[from] Box<ProtoArchiveError>),
+
+    #[diagnostic(transparent)]
+    #[error(transparent)]
+    Client(#[from] Box<WarpgateClientError>),
+
+    #[diagnostic(transparent)]
+    #[error(transparent)]
     Config(#[from] Box<ProtoConfigError>),
+
+    #[diagnostic(transparent)]
+    #[error(transparent)]
+    Fs(#[from] Box<FsError>),
 
     #[diagnostic(transparent)]
     #[error(transparent)]
@@ -22,7 +37,7 @@ pub enum ProtoResolveError {
 
     #[diagnostic(transparent)]
     #[error(transparent)]
-    Tool(#[from] Box<ProtoToolError>),
+    Process(#[from] Box<ProtoProcessError>),
 
     #[diagnostic(code(proto::resolve::offline::version_required))]
     #[error(
@@ -33,7 +48,7 @@ pub enum ProtoResolveError {
     RequiredInternetConnectionForVersion { command: String, bin_dir: PathBuf },
 
     #[diagnostic(code(proto::resolve::invalid_version))]
-    #[error("Invalid version or requirement {}.", .version.style(Style::Hash))]
+    #[error("Invalid version or requirement in tool specification {}.", .version.style(Style::Hash))]
     InvalidVersionSpec {
         version: String,
         #[source]
@@ -59,11 +74,40 @@ pub enum ProtoResolveError {
         .version.style(Style::Hash),
     )]
     FailedVersionResolve { tool: String, version: String },
+
+    #[diagnostic(code(proto::resolve::unknown_backend))]
+    #[error(
+        "Unknown backend in tool specification {}. Only {} are supported.",
+        .spec.style(Style::Hash),
+        .backends.iter().map(|be| be.to_string().style(Style::Id)).collect::<Vec<_>>().join(", ")
+    )]
+    UnknownBackend {
+        backends: Vec<Backend>,
+        spec: String,
+    },
+}
+
+impl From<ProtoArchiveError> for ProtoResolveError {
+    fn from(e: ProtoArchiveError) -> ProtoResolveError {
+        ProtoResolveError::Archive(Box::new(e))
+    }
+}
+
+impl From<WarpgateClientError> for ProtoResolveError {
+    fn from(e: WarpgateClientError) -> ProtoResolveError {
+        ProtoResolveError::Client(Box::new(e))
+    }
 }
 
 impl From<ProtoConfigError> for ProtoResolveError {
     fn from(e: ProtoConfigError) -> ProtoResolveError {
         ProtoResolveError::Config(Box::new(e))
+    }
+}
+
+impl From<FsError> for ProtoResolveError {
+    fn from(e: FsError) -> ProtoResolveError {
+        ProtoResolveError::Fs(Box::new(e))
     }
 }
 
@@ -79,8 +123,8 @@ impl From<WarpgatePluginError> for ProtoResolveError {
     }
 }
 
-impl From<ProtoToolError> for ProtoResolveError {
-    fn from(e: ProtoToolError) -> ProtoResolveError {
-        ProtoResolveError::Tool(Box::new(e))
+impl From<ProtoProcessError> for ProtoResolveError {
+    fn from(e: ProtoProcessError) -> ProtoResolveError {
+        ProtoResolveError::Process(Box::new(e))
     }
 }
