+++ crates/core/src/lib.rs
@@ -6,11 +6,12 @@ mod env_error;
 pub mod flow;
 mod helpers;
 pub mod layout;
+mod loader;
+mod loader_error;
 mod lockfile;
 pub mod registry;
 mod tool;
 mod tool_error;
-mod tool_loader;
 mod tool_manifest;
 mod tool_spec;
 pub mod utils;
@@ -21,10 +22,11 @@ pub use config_error::*;
 pub use env::*;
 pub use env_error::*;
 pub use helpers::*;
+pub use loader::*;
+pub use loader_error::*;
 pub use lockfile::*;
 pub use tool::*;
 pub use tool_error::*;
-pub use tool_loader::*;
 pub use tool_manifest::*;
 pub use tool_spec::*;
 pub use version_resolver::*;
