+++ crates/core/src/loader.rs
@@ -1,7 +1,7 @@
 use crate::config::SCHEMA_PLUGIN_KEY;
 use crate::env::ProtoEnvironment;
+use crate::loader_error::ProtoLoaderError;
 use crate::tool::Tool;
-use crate::tool_error::ProtoToolError;
 use crate::tool_spec::Backend;
 use convert_case::{Case, Casing};
 use rustc_hash::FxHashSet;
@@ -16,7 +16,7 @@ pub fn inject_proto_manifest_config(
     id: &Id,
     proto: &ProtoEnvironment,
     manifest: &mut PluginManifest,
-) -> Result<(), ProtoToolError> {
+) -> Result<(), ProtoLoaderError> {
     let config = proto.load_config()?;
 
     if let Some(tool_config) = config.tools.get(id) {
@@ -33,7 +33,7 @@ pub fn inject_proto_manifest_config(
 }
 
 #[instrument(skip(proto))]
-pub fn locate_tool(id: &Id, proto: &ProtoEnvironment) -> Result<PluginLocator, ProtoToolError> {
+pub fn locate_tool(id: &Id, proto: &ProtoEnvironment) -> Result<PluginLocator, ProtoLoaderError> {
     let mut locator = None;
     let configs = proto.load_config_manager()?;
 
@@ -66,7 +66,7 @@ pub fn locate_tool(id: &Id, proto: &ProtoEnvironment) -> Result<PluginLocator, P
     }
 
     let Some(locator) = locator else {
-        return Err(ProtoToolError::UnknownTool { id: id.to_owned() });
+        return Err(ProtoLoaderError::UnknownTool { id: id.to_owned() });
     };
 
     Ok(locator)
@@ -74,7 +74,7 @@ pub fn locate_tool(id: &Id, proto: &ProtoEnvironment) -> Result<PluginLocator, P
 
 pub async fn load_schema_plugin_with_proto(
     proto: impl AsRef<ProtoEnvironment>,
-) -> Result<PathBuf, ProtoToolError> {
+) -> Result<PathBuf, ProtoLoaderError> {
     let proto = proto.as_ref();
     let schema_id = Id::raw(SCHEMA_PLUGIN_KEY);
     let schema_locator = locate_tool(&schema_id, proto)?;
@@ -87,7 +87,7 @@ pub async fn load_schema_plugin_with_proto(
     Ok(path)
 }
 
-pub fn load_schema_config(plugin_path: &Path) -> Result<json::JsonValue, ProtoToolError> {
+pub fn load_schema_config(plugin_path: &Path) -> Result<json::JsonValue, ProtoLoaderError> {
     let mut is_toml = false;
     let mut schema: json::JsonValue = match plugin_path.extension().and_then(|ext| ext.to_str()) {
         Some("toml") => {
@@ -158,7 +158,7 @@ pub async fn load_tool_from_locator(
     id: impl AsRef<Id> + Debug,
     proto: impl AsRef<ProtoEnvironment>,
     locator: impl AsRef<PluginLocator> + Debug,
-) -> Result<Tool, ProtoToolError> {
+) -> Result<Tool, ProtoLoaderError> {
     let id = id.as_ref();
     let proto = proto.as_ref();
     let locator = locator.as_ref();
@@ -204,7 +204,7 @@ pub async fn load_tool(
     id: &Id,
     proto: &ProtoEnvironment,
     mut backend: Option<Backend>,
-) -> Result<Tool, ProtoToolError> {
+) -> Result<Tool, ProtoLoaderError> {
     // Determine the backend plugin to use
     if backend.is_none() {
         let config = proto.load_config()?;
